lint: fix clippy module inception warning
